requirements.txt
<<<<<<< HEAD
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
boto3==1.42.44
botocore==1.42.44
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
ImageHash==4.3.2
jmespath==1.1.0
keyboard==0.13.5
MouseInfo==0.1.3
multidict==6.7.1
numpy==2.4.2
pillow==12.1.0
propcache==0.4.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytweening==1.2.0
PyWavelets==1.9.0
s3transfer==0.16.0
scipy==1.17.0
six==1.17.0
urllib3==2.6.3
yarl==1.22.0
=======
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
asyncpg==0.31.0
attrs==25.4.0
boto3==1.42.42
botocore==1.42.42
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
discord.py==2.6.4
dotenv==0.9.9
et_xmlfile==2.0.0
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
idna==3.11
ImageHash==4.3.2
jmespath==1.1.0
lxml==6.0.2
multidict==6.7.1
numpy==2.4.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==12.1.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyWavelets==1.9.0
pywin32-ctypes==0.2.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.16.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
xlrd==2.0.2
yarl==1.22.0
>>>>>>> fba511dfb6b5229b5842777eae7386494812dd68
